requirements.txt
gradio==5.0.0
groq==1.0.0
sentence-transformers==3.3.1
faiss-cpu==1.10.1
pymupdf==1.24.10
python-dotenv==1.1.0
langchain==0.3.0
langchain-community==0.3.0